feat(items): update the needle and bloody needle models
