+++ src/main/java/org/sosly/witchcraft/items/ItemRegistry.java
@@ -12,6 +12,7 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 import org.sosly.witchcraft.Witchcraft;
 import org.sosly.witchcraft.items.armor.CovenArmorItem;
+import org.sosly.witchcraft.items.sympathy.BloodyNeedleItem;
 
 @Mod.EventBusSubscriber(modid= Witchcraft.MOD_ID, bus= Mod.EventBusSubscriber.Bus.MOD)
 public class ItemRegistry {
@@ -24,6 +25,7 @@ public class ItemRegistry {
             MAArmorMaterial.INFUSED_SILK, ArmorItem.Type.HELMET, new Item.Properties()));
     public static final RegistryObject<Item> COVEN_ARMOR_BOOTS = ITEMS.register("coven_armor_boots", () -> new CovenArmorItem(
             MAArmorMaterial.INFUSED_SILK, ArmorItem.Type.BOOTS, new Item.Properties()));
+    public static final RegistryObject<Item> BLOODY_NEEDLE = ITEMS.register("bloody_needle", BloodyNeedleItem::new);
 
     @SubscribeEvent
     public static void FillCreativeTabs(BuildCreativeModeTabContentsEvent event) {
