+++ src/main/resources/assets/mna/models/item/vinteum_needle.json
@@ -1,0 +1,29 @@
+{
+  "replace": true,
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "mna:item/artifice/vinteum_needle"
+  },
+  "display": {
+    "thirdperson_righthand": {
+      "rotation": [0, 90, 180],
+      "translation": [0, 0, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "firstperson_righthand": {
+      "rotation": [0, 90, 180],
+      "translation": [0, 5, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "thirdperson_lefthand": {
+      "rotation": [90, 90, 0],
+      "translation": [0, 0, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "firstperson_lefthand": {
+      "rotation": [90, 90, 0],
+      "translation": [0, 5, 0],
+      "scale": [0.25, 0.25, 0.25]
+    }
+  }
+}
