+++ src/main/resources/assets/mnaw/models/item/bloody_needle.json
@@ -1,0 +1,28 @@
+{
+  "parent": "item/generated",
+  "textures": {
+    "layer0": "mnaw:item/bloody_needle"
+  },
+  "display": {
+    "thirdperson_righthand": {
+      "rotation": [0, 90, 180],
+      "translation": [0, 0, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "firstperson_righthand": {
+      "rotation": [0, 90, 180],
+      "translation": [0, 5, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "thirdperson_lefthand": {
+      "rotation": [90, 90, 0],
+      "translation": [0, 0, 0],
+      "scale": [0.25, 0.25, 0.25]
+    },
+    "firstperson_lefthand": {
+      "rotation": [90, 90, 0],
+      "translation": [0, 5, 0],
+      "scale": [0.25, 0.25, 0.25]
+    }
+  }
+}
