+++ src/main/resources/assets/mnaw/models/item/poppet.json
@@ -3,7 +3,7 @@
   "display": {
     "gui": {
       "translation": [0, -0.1, 0],
-      "scale": [0.75, 0.75, 0.75],
+      "scale": [0.5, 0.5, 0.5],
       "rotation": [-0.1, -0.1, 0]
     }
   }
