+++ src/main/resources/assets/mnaw/textures/item/bloody_needle.png.mcmeta
@@ -1,0 +1,7 @@
+{
+    "animation": {
+        "interpolate": true,
+        "frametime": 20,
+        "frames": [0,1,2,3]
+    }
+}
