+++ src/main/resources/data/mnaw/recipes/poppet.json
@@ -12,5 +12,8 @@
     "A": {
         "item": "mna:animus_dust"
     }
+  },
+  "result": {
+    "item": "mnaw:poppet"
   }
 }
